requirements.txt
python-dotenv
requests
pyairtable
apify-client
elevenlabs
openai
pytz
hermes-agent
elevenlabs[conversational_ai]
pyaudio
Flask